WIP generate global tsdf, but visulization tsdf server node update really slow

+++ coxgraph/include/coxgraph/common.h
@@ -18,7 +18,7 @@
 
 namespace coxgraph {
 
-typedef int8_t CliId;
+typedef int16_t CliId;
 
 using CliSm = voxgraph::VoxgraphSubmap;
 using SerSmId = voxgraph::SubmapID;
